+++ src/main/java/insane96mcp/insanelib/InsaneLib.java
@@ -7,6 +7,7 @@ import com.mojang.logging.LogUtils;
 import insane96mcp.insanelib.command.ILCommand;
 import insane96mcp.insanelib.data.AttributeModifierOperationSerializer;
 import insane96mcp.insanelib.data.JsonFeatureDataReloadListener;
+import insane96mcp.insanelib.data.poolinjection.PoolInjectionReloadListener;
 import insane96mcp.insanelib.datagen.ILItemTagProvider;
 import insane96mcp.insanelib.module.base.PushResistance;
 import insane96mcp.insanelib.module.base.items.ItemComponentsReloadListener;
@@ -69,6 +70,7 @@ public class InsaneLib {
         NeoForge.EVENT_BUS.addListener(this::onAddReloadListeners);
         NeoForge.EVENT_BUS.addListener(this::onRegisterCommands);
         NeoForge.EVENT_BUS.addListener(ILCriteriaTriggers::onBlockBreak);
+        NeoForge.EVENT_BUS.addListener(PoolInjectionReloadListener::onLootTableLoad);
     }
 
     private void onRegisterCommands(RegisterCommandsEvent event) {
@@ -78,6 +80,7 @@ public class InsaneLib {
     private void onAddReloadListeners(AddReloadListenerEvent event) {
         event.addListener(JsonFeatureDataReloadListener.INSTANCE);
         event.addListener(ItemComponentsReloadListener.INSTANCE);
+        event.addListener(PoolInjectionReloadListener.INSTANCE);
     }
 
     @SubscribeEvent